Creating assert functions
This set of three functions can be used to add runtime assertions to
code.

The assertions are all gated by a constant ShouldAssert boolean. This
constant is 'false' by default, but the build tag '-tags runssert' sets
this constant to be true and allows the runtime assert to execute.

We chose a constant value to gate this behaviour because it allows the
assert functions to be completely eliminated by the compiler when the
asserts are disabled.